requirements_kato.txt
absl-py==1.0.0
aiofiles==23.2.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
astunparse==1.6.3
attrs==23.2.0
backports.functools-lru-cache==1.6.4
blosc2==2.5.1
cachetools==5.3.3
carbontracker==1.2.5
certifi==2017.4.17
cffi==1.16.0
charset-normalizer==2.0.12
click==8.1.7
convertdate==2.3.2
coverage==4.0.3
cycler==0.11.0
Cython==0.29.27
dataclasses==0.6
decorator==4.4.2
diffusers==0.29.1
distlib==0.3.8
dnspython==2.6.1
docopt==0.6.2
einops==0.8.0
email_validator==2.2.0
entrypoints==0.4
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
ffmpy==0.3.2
filelock==3.14.0
flatbuffers==24.3.25
fsspec==2024.6.0
future==0.18.2
gast==0.4.0
geocoder==1.38.1
gluonts==0.8.1
google-auth==2.29.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gradio==4.37.1
gradio_client==1.0.2
grpcio==1.35.0
h11==0.14.0
h5py==3.11.0
hdbscan==0.8.37
hijri-converter==2.2.3
hmmlearn==0.2.7
holidays==0.13
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.3
imageio==2.15.0
imageio-ffmpeg==0.4.5
importlib-metadata==4.2.0
importlib_resources==6.4.0
ipykernel==5.5.5
ipython==5.8.0
ipython-genutils==0.2.0
Jinja2==3.1.4
joblib==1.1.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-client==5.3.4
jupyter_core==5.0.0
keras==2.11.0
kiwisolver==1.3.1
kneed==0.8.5
korean-lunar-calendar==0.2.1
libclang==18.1.1
llvmlite==0.43.0
Markdown==3.3.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.4.3
mdurl==0.1.2
mne==1.5.1
more-itertools==8.12.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.8
ndindex==1.8
networkx==3.2.1
nose==1.3.7
numba==0.60.0
numexpr==2.10.0
numpy==1.23.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python==4.10.0.84
opt-einsum==3.3.0
orjson==3.10.5
packaging==24.0
pandas==1.5.3
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pipreqs==0.4.11
platformdirs==4.2.2
pooch==1.8.1
proglog==0.1.9
prompt-toolkit==1.0.15
protobuf==3.19.4
psg-utils==0.1.6
psutil==5.9.8
ptyprocess==0.7.0
py-cpuinfo==9.0.0
pyasn1==0.4.8
pyasn1_modules==0.4.0
pybasicbayes==0.2.2
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
pydub==0.25.1
pyEDFlib==0.1.28
Pygments==2.18.0
PyMeeus==0.5.11
pynndescent==0.5.13
pynvml==11.5.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
ratelim==0.1.6
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
ruff==0.4.10
safetensors==0.4.3
scikit-learn==1.1.3
scipy==1.11.4
seaborn==0.11.2
semantic-version==2.10.0
shellingham==1.5.4
simplegeneric==0.8.1
six==1.16.0
sklearn==0.0
sniffio==1.3.1
soundfile==0.12.1
starlette==0.37.2
statsmodels==0.12.2
sympy==1.12.1
tables==3.9.2
tabulate==0.9.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.1
tensorflow-addons==0.23.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.37.0
termcolor==2.4.0
threadpoolctl==3.1.0
tomlkit==0.12.0
toolz==0.11.2
torch==2.1.2+cu118
torch-tb-profiler==0.3.1
torchaudio==2.1.2+cu118
torchvision==0.16.2+cu118
tornado==6.1
tqdm==4.62.3
traitlets==4.3.3
triton==2.1.0
typeguard==2.13.3
typer==0.12.3
typing_extensions==4.12.1
ujson==5.10.0
umap-learn==0.5.6
urllib3==2.2.2
utime==1.0.0
uvicorn==0.30.1
uvloop==0.19.0
virtualenv==20.26.2
watchfiles==0.22.0
wcwidth==0.2.5
websockets==11.0.3
Werkzeug==1.0.1
wfdb==4.1.2
wrapt==1.16.0
yamlhparams==0.2.0
yarg==0.1.9
zipp==3.19.2
